Add query for vmSettings
