+++ tests/protocol/test_wire.py
@@ -31,7 +31,7 @@ from azurelinuxagent.common.agent_supported_feature import SupportedFeatureNames
     get_agent_supported_features_list_for_crp
 from azurelinuxagent.common.exception import ResourceGoneError, ProtocolError, \
     ExtensionDownloadError, HttpError
-from azurelinuxagent.common.protocol.goal_state import ExtensionsConfig, GoalState
+from azurelinuxagent.common.protocol.goal_state import ExtensionsConfig
 from azurelinuxagent.common.protocol.hostplugin import HostPluginProtocol
 from azurelinuxagent.common.protocol.restapi import VMAgentManifestUri
 from azurelinuxagent.common.protocol.wire import WireProtocol, WireClient, \
@@ -39,7 +39,6 @@ from azurelinuxagent.common.protocol.wire import WireProtocol, WireClient, \
 from azurelinuxagent.common.telemetryevent import GuestAgentExtensionEventsSchema, \
     TelemetryEventParam, TelemetryEvent
 from azurelinuxagent.common.utils import restutil
-from azurelinuxagent.common.exception import IncompleteGoalStateError
 from azurelinuxagent.common.version import CURRENT_VERSION, DISTRO_NAME, DISTRO_VERSION
 from azurelinuxagent.ga.exthandlers import get_exthandlers_handler
 from tests.ga.test_monitor import random_generator
@@ -1203,123 +1202,6 @@ class UpdateGoalStateTestCase(AgentTestCase):
                     len(matches),
                     len(protected_settings),
                     "Could not find the expected number of redacted settings. Expected {0}.\n{1}".format(len(protected_settings), extensions_config))
-
-
-class TryUpdateGoalStateTestCase(HttpRequestPredicates, AgentTestCase):
-    """
-    Tests for WireClient.try_update_goal_state()
-    """
-    def test_it_should_return_true_on_success(self):
-        with mock_wire_protocol(mockwiredata.DATA_FILE) as protocol:
-            self.assertTrue(protocol.client.try_update_goal_state(), "try_update_goal_state should have succeeded")
-
-    def test_incomplete_gs_should_fail(self):
-
-        with mock_wire_protocol(mockwiredata.DATA_FILE) as protocol:
-            GoalState.fetch_full_goal_state(protocol.client)
-
-            protocol.mock_wire_data.data_files = mockwiredata.DATA_FILE_NOOP_GS
-            protocol.mock_wire_data.reload()
-            protocol.mock_wire_data.set_incarnation(2)
-
-            with self.assertRaises(IncompleteGoalStateError):
-                GoalState.fetch_full_goal_state_if_incarnation_different_than(protocol.client, 1)
-
-    def test_it_should_return_false_on_failure(self):
-        with mock_wire_protocol(mockwiredata.DATA_FILE) as protocol:
-            def http_get_handler(url, *_, **__):
-                if self.is_goal_state_request(url):
-                    return HttpError('Exception to fake an error retrieving the goal state')
-                return None
-
-            protocol.set_http_handlers(http_get_handler=http_get_handler)
-
-            self.assertFalse(protocol.client.try_update_goal_state(), "try_update_goal_state should have failed")
-
-    def test_it_should_log_errors_only_when_the_error_state_changes(self):
-        with mock_wire_protocol(mockwiredata.DATA_FILE) as protocol:
-            def http_get_handler(url, *_, **__):
-                if self.is_goal_state_request(url):
-                    if fail_goal_state_request:
-                        return HttpError('Exception to fake an error retrieving the goal state')
-                return None
-
-            protocol.set_http_handlers(http_get_handler=http_get_handler)
-
-            @contextlib.contextmanager
-            def create_log_and_telemetry_mocks():
-                with patch("azurelinuxagent.common.protocol.wire.logger", autospec=True) as logger_patcher:
-                    with patch("azurelinuxagent.common.protocol.wire.add_event") as add_event_patcher:
-                        yield logger_patcher, add_event_patcher
-
-            calls_to_strings = lambda calls: (str(c) for c in calls)
-            filter_calls = lambda calls, regex=None: (c for c in calls_to_strings(calls) if regex is None or re.match(regex, c))
-            logger_calls = lambda regex=None: [m for m in filter_calls(logger.method_calls, regex)]  # pylint: disable=used-before-assignment,unnecessary-comprehension
-            warnings = lambda: logger_calls(r'call.warn\(.*An error occurred while retrieving the goal state.*')
-            periodic_warnings = lambda: logger_calls(r'call.periodic_warn\(.*Attempts to retrieve the goal state are failing.*')
-            success_messages = lambda: logger_calls(r'call.info\(.*Retrieving the goal state recovered from previous errors.*')
-            telemetry_calls = lambda regex=None: [m for m in filter_calls(add_event.mock_calls, regex)]  # pylint: disable=used-before-assignment,unnecessary-comprehension
-            goal_state_events = lambda: telemetry_calls(r".*op='FetchGoalState'.*")
-
-            #
-            # Initially calls to retrieve the goal state are successful...
-            #
-            fail_goal_state_request = False
-            with create_log_and_telemetry_mocks() as (logger, add_event):
-                protocol.client.try_update_goal_state()
-
-                lc = logger_calls()
-                self.assertTrue(len(lc) == 0, "A successful call should not produce any log messages: [{0}]".format(lc))
-
-                tc = telemetry_calls()
-                self.assertTrue(len(tc) == 0, "A successful call should not produce any telemetry events: [{0}]".format(tc))
-
-            #
-            # ... then an error happens...
-            #
-            fail_goal_state_request = True
-            with create_log_and_telemetry_mocks() as (logger, add_event):
-                protocol.client.try_update_goal_state()
-
-                w = warnings()
-                pw = periodic_warnings()
-                self.assertEqual(len(w), 1, "A failure should have produced a warning: [{0}]".format(w))
-                self.assertEqual(len(pw), 1, "A failure should have produced a periodic warning: [{0}]".format(pw))
-
-                gs = goal_state_events()
-                self.assertTrue(len(gs) == 1 and 'is_success=False' in gs[0], "A failure should produce a telemetry event (success=false): [{0}]".format(gs))
-
-            #
-            # ... and errors continue happening...
-            #
-            with create_log_and_telemetry_mocks() as (logger, add_event):
-                protocol.client.try_update_goal_state()
-                protocol.client.try_update_goal_state()
-                protocol.client.try_update_goal_state()
-
-                w = warnings()
-                pw = periodic_warnings()
-                self.assertTrue(len(w) == 0, "Subsequent failures should not produce warnings: [{0}]".format(w))
-                self.assertEqual(len(pw), 3, "Subsequent failures should produce periodic warnings: [{0}]".format(pw))
-
-                tc = telemetry_calls()
-                self.assertTrue(len(tc) == 0, "Subsequent failures should not produce any telemetry events: [{0}]".format(tc))
-
-            #
-            # ... until we finally succeed
-            #
-            fail_goal_state_request = False
-            with create_log_and_telemetry_mocks() as (logger, add_event):
-                protocol.client.try_update_goal_state()
-
-                s = success_messages()
-                w = warnings()
-                pw = periodic_warnings()
-                self.assertEqual(len(s), 1, "Recovering after failures should have produced an info message: [{0}]".format(s))
-                self.assertTrue(len(w) == 0 and len(pw) == 0, "Recovering after failures should have not produced any warnings: [{0}] [{1}]".format(w, pw))
-
-                gs = goal_state_events()
-                self.assertTrue(len(gs) == 1 and 'is_success=True' in gs[0], "Recovering after failures should produce a telemetry event (success=true): [{0}]".format(gs))
 
 
 class UpdateHostPluginFromGoalStateTestCase(AgentTestCase):
